gamification-service/requirements.txt
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
itsdangerous==2.2.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.6.1
PyJWT==2.10.1
redis>=5.0.0,<6
httpx==0.28.1
alembic==1.14.0
psycopg2-binary==2.9.10
